Directional Lighting, Code Organization, Debug Shader, etc.
Many changes...
-Added directional lighting which involves computing the normal matrix
which is the inverse transpose of the view matrix.
-Consolidated architecture by making the SceneManager contain a
ViewManager which contains the SimpleContext.  SceneManager also
contains a ModelFactory which acts as a singleton that generates models.
Models contain shaders and these shaders operate on a model object when
running the shader.  Streamlined event update management.
-Adding a debug shader for visualizing normals.  This helped with
proving the normals were accurate and transformed correctly by the
normal matrix.
-Added a new design picture of how everything fits where.