Cache Animations to GPU
Added a capability to choose between caching the animation frames to the
GPU or streaming buffer data from RAM to the GPU dynamically.  The
latter is slower and is more practical because it does not allocate
nearly as much memory as the cached animation scheme does.  Animations
default to streaming buffer data dynamically from RAM to GPU.