GPU-side animation computing
1. Original animation computing was processed on the CPU side and has
now been converted to be processed on the GPU using weights and indexes
as attributes and bone matrices as a uniform matrix array.  The bone
matrices change every animation frame while the indexes are used to
index into the bone that affects each vertex which is accompanied with a
scalar weight.  As of now there can only be up to four bone matrices
that can affect one vertex so the index and weight attributes add 8
floats to each vertex shader.
2.  Texture broker is now a static texture manager that lives in the
model class which allows reuse of textures that are already loaded to
gpu texture memory.  Just pass around texture contexts to models.
3.  Removed double buffering because the GPU-side animation computations
now allow up to 600 animated entities running at 120 fps.  The original
CPU-side computations allowed up to 100 animated entities running at 60
fps with many optimizations in place.  This is a huge speed
improvement!!!!!!!!!!!!!!!!!!!!!!!!