Multi-texture models
Implemented the capability to have multiple textures used in one model.
The idea is that all of the vertex, normal and texture UV data is each a
single large buffer regardless of texture changes.  Meta data is stored
notifying the shader to switch textures based on vertex stride
information.  The shader program does not need to be restarted every
texture change, the shader manager only needs to change the active
texture and the starting index in the attribute buffer along with a
length...very nice!