Octree Space Partitioning
An octree space partitioning algorithm was used to prevent having more
than N primitives in a certain subspace to test for collisions.  The
octree grows until each node on the tree does not contain more than N
primitives and the nodes that do not have children (leaves) are used to
test for collisions.  Each simulation step needs to update if primitives
have moved to or left other nodes in the tree and then all of the
primitives within a node are testing against each other as long as one
of the primitives is in motion.  The algorithm improves collision
detection speeds drastically.