View Controller Modes
1) The view controller has two modes in which the user's view can be
manipulated.  The first mode is called "God mode" and can be entered by
pressing 'g', which essentially allows the view to be changed using wasd
and not be affected by kinematics.  The second mode is called model
viewer mode where the user's view changes to the third person view of a
model in a scene by pressing 'q' over and over which will cycle through
all the models.
2) Improved the efficiency of the OSP update call by searching from the
octree root node for a new subspacec location of a primitive.  Before I
used all of the leaves (end nodes) to search for updates in position
which is inefficient but got the job done prior.