Force driven mechanics
1) A class called FunctionState is initialized with a lambda function
describing the equation that you could model for a physics based
simulation.  A function pointer is another important parameter that gets
invoked at a certain rate millisecond rate.  This class acts as a
background worker that sends physics update information based on user
input.  The first use case is the view manager controlling a model and
adding forces to the model which then computes a step function
periodically and affects the model kinematic state with function pointer
updates.
2) OSP optimizations which still need work.
3) Pass by reference performance fixes.
4) Fixed openGL barf when calling glDrawArraysEXT, changed to plain
glDrawArrays.
5) Input model is based on a key press and release not on repeating
keystrokes anymore which is now disabled.