Generic skybox
Discard any fragment in the deferred rendering pass if the position is
(0,0,0).  This disables any data being written to the depth or color
buffer.  If a fragment passes, then write a depth value of 0.  A black
and white depth buffer is written indicating fragment data or no
fragment data.  In the skybox pass I still keep depth testing in
normalized device coordinates which results in a depth buffer of all 0.5
value.  0.5 is less than 1.0 which passes the depth test for the skybox
fragments when the depth buffer is 1.0 from zero fragment data.