Improved shader uniform handling
Created a uniform class that gathers all the uniforms in a shader and
stores them in the shader class object.  A uniform function takes in a
void pointer and gets cast to the actual type using the uniform info
stored in the shader object!  Still need to change the compute shaders
to this paradigm.