Fixed camera issues
The major issue was a race condition in which a state vector was being
written to by multiple threads. This would result in negating the view
matrix which caused jumping to different axis regions and causing
violent velocity buffers lol.  Added lock guards in the state vector
class which needed to be done a long time ago.