Dynamic model loading
Model class has been abstracted under the ModelBroker which builds all
models and distributes them to Entity classes that require them.  The
command "update <model_name>" has been added to the terminal which
allows dynamic reloading of a model allowing an artist to change a model
then just reload it in the engine.  Fixed some uniform issues associated
with too small of a buffer.
Directional shadows are being reworked and thus do not function at the
moment.....