ViewManager refactor
Added a camera class that allows viewing of multiple camera frustum for
eventual frustum occlusion, etc.  For now the frustum are not visible
because the alpha value in the shader is set to 0.  To enable frustum
visualizers just set the alpha value to 0.2 in the debugShader.frag
during runtime using the recompile command and they will magically
appear for your debugging pleasure!  Still need to clean up the state
vector part of the view manager class to make it cleaner.