Mouse picking using id buffer
A function callback executes when a left mouse click event has occurred.
The id g buffer texture is read back from the GPU and the mouse click
coordinates index into the texture buffer.  Then we know which id value
is for a particular entity and we can highlight via AABB is the next
step!