Entity AABBs are shown when an item is selected
Neat feature that I will use to move characters in space, etc.  Need a
better way to approximate animated object's AABBs because the skinning
gets done on the GPU which gets more accurate min max point values for
the AABB calculation.  Now they are super sloppy but the picking is
pixel perfect!