Flexible resolution and tiling texture fix
Picking with any resolution is now supported.  The tiling texture fix
needed to add an offset to the primitive id.  This was necessary because
some vao have different textures and thus have multiple draw calls per
entity ID.  The simple fix was to keep track of all the primitives
rendered per vao and offset it from the built in primitive_id when
rendering the id texture.  TODO: There is a slight discontinutiy when
texturing between tiles that needs to be smoothly interpolated.