Removed model dependencies
Meshes have two categories that are static and scene.  Static meshes are
typically single models that would include a tree, bush, person or
house.  Scene meshes contain many instances of static meshes and are
therefore dependent on static meshes to be loaded prior to loading a
scene mesh.  The model loader loads all of the static meshes prior to
scene meshes now.  A scene mesh loads all of the instances of models as
entities which is different than how a static mesh is loaded.