Fire and smoke effects for DX12
Needed to change prim topology for fullscreen single tri and quad tri
strip for billboarding.  Both get treated essentially the same in the
shader because they only require a vertex id from an index buffer and
the draw command dicatates whether 3 or 4 verts are passed to the input
assembler.  I'm sure this will cause shit to break in the future lol.