Added bloom effects to DX12.
A bunch of compute shaders were brought in to support bloom effects.
Compute barriers are added in quite a lot of places now.  I am sure
there is a bug or two in this first pass implementation but that will be
for later.  Great progress.  Almost have full support crossover from
openGL at this point.