Layered texture reuse
Layered textures were not reusing already resident sub textures that
were created on the gpu previously by other layered textures.  This was
causing a large amount of memory to be consumed when a large number of
tiles are in a scene.  Memory usage is no longer an issue and the load
time is less than 8 seconds for a scene that used to take 30 seconds.