Fixed picking for left handed coordinate system
First issue is that when you read the id buffer off the frame buffer you
are reading back the id buffers in the right handed coordinate system of
opengl.  I need to transform the triangles back to the left handed
coordinate system to correctly fetcht the triangles.  I no longer load
the mesh data into the scene file; instead we instance off of each model
so the scene files are tiny and only contain string references pointing
to the static models.