Temporary fix to unify opengl and directx
Coordinate system changes are fixed and it was because i didn't realize
that the view matrix only needed to be transformed when changing
coordinate systems.  I now only convert the view matrix and leave the
model-world space as is.  The view matrix is in charge of transforming
to the particular camera space's coordinate system!