Fixed texture and pathing algorithms
Each call to edit the terrain texture was allocating a whole texture block of memory so i fixed that.  The pathing algorithm starts in the middle of the tild grid and flows out until it hit the end of the tile grid.  Still needs more fine tuning because it doesn't look like a path at all...also I need to place the trees after the path is done in order to not have trees in the middle of paths, etc...