Fixed pathing issues and added random rotation to trees.
Normals are broke to shit when adding a rotation transform to them in object space.  This needs to be fixed in the StaticShader.cpp where we pass in the normal matrix.  The normal matrix needs to be the view's normal matrix and the local rotation transform combined but that doesn't seem to be working at the moment.  The pathing error was in the tree indexing which was using width instead of length.  I think there is still a small problem with pathing but it rarely shows up......ahhhhh it never ends lol