Fixed item placement
The issue is that item placement is at a different granularity than the path texturing so in order to make sure trees aren't placed on the path I need to write the path index to the neighbors.  This adds another issue where creating paths seem to terminate right away sometimes so i need to look into that as well.