Pathing bug fixes
Path ids were not being set correctly and paths were being written to incorrect tiles because the check to make sure a path is in bounds was before the incremental change of the path.  These changes fixed a lot of random bugs that were happening including early termination of a parent path after spawning a new path.  It's good to be back and breaks are necessary.