corruption fix for normals
Leaves have overlapping triangles with diffenet windings indicating front and back face.  Essentially the fix was to cull back facing triangles and manually select the mip level to 0 for now.  The mip level will be calculated by taking the ray vector and the normal vector of the intersected triangle and doing a dot product.  This will give a value between 1 and 0 that will be scaled up to mip 7 and mip 0.